Commit Automatico do Vim - Arch Linux

+++ TerceiroSemestre/Beecrowd02.c
@@ -1,18 +1,18 @@
 #include<stdio.h>
 
 int main() {
-   float N;
+   double N;
    scanf("%f", &N);
    
    if (N <= 25 && N >= 0) {
        printf("intervalo [0,25]\n");
-   } else if (50.0000000 >= N && N > 25.0000000) {
+   } else if (50 >= N && N > 25) {
        printf("intervalo (25,50]");
-   } else if (75.0000000 >= N && N > 50.0000000) {
+   } else if (75 >= N && N > 50) {
        printf("intervalo (50,75]");
-   } else if (100.0000000 >= N && N > 75.0000000) {
+   } else if (100 >= N && N > 75) {
        printf("intervalo (75,100]");
-   } else if (N > 100.0000000 && N <= 1) {
+   } else if (N > 100 && N <= 1) {
        printf("Fora de intervalo");
    } else {
        printf("Fora de intervalo");
